vexa-core/requirements.txt
fastapi
uvicorn
pydantic
requests
pytest
slack-bolt
python-dotenv
